fix : jwt 유효시간 타입 버그 수정

+++ src/main/java/com/coniverse/dangjang/domain/auth/service/AuthTokensGenerator.java
@@ -19,8 +19,8 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class AuthTokensGenerator {
 	private static final String BEARER_TYPE = "Bearer";
-	private static final long ACCESS_TOKEN_EXPIRE_TIME = 1000 * 60 * 30;            // 30분
-	private static final long REFRESH_TOKEN_EXPIRE_TIME = 1000 * 60 * 60 * 24 * 7;  // 7일
+	private static final long ACCESS_TOKEN_EXPIRE_TIME = (long)1000 * 60 * 30;            // 30분
+	private static final long REFRESH_TOKEN_EXPIRE_TIME = (long)1000 * 60 * 60 * 24 * 7;  // 7일
 
 	private final JwtTokenProvider jwtTokenProvider;
